requirements.txt
annotated-types==0.6.0
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.3.1
Django==4.2.6
django-bootstrap-datepicker-plus==5.0.4
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
idna==3.4
pydantic==2.4.2
pydantic_core==2.10.1
PyJWT==2.8.0
python-decouple==3.8
pytz==2023.3.post1
requests==2.31.0
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
